Fix calculation of the node dependencies of an artifact transform that takes upstream dependencies as input, when those upstream dependencies include an external dependency that conflicts with a project.
The basic problem is that the full set of node dependencies for a transform are not known until the entire artifact dependency graph has been resolved and all conflicts resolved. However, this does not always happen at work graph calculation time and some node dependencies are "discovered" at execution time.

In this commit, use the pre- and post-execution node concepts to allow the "finalize upstream dependencies for transform" node to delay the transform until all node dependencies of the transform have been discovered and completed.