Detect wrapper configuration changes between TAPI build invocations
Fixes #21137

IDEA users reported that builds sometimes are running with an unexpected Gradle
version. A typical reproducer is to:

- check out a project from git and import it into IDEA
- run a task using the Gradle wrapper
- switch to a different branch that uses a different wrapper version
- run the same build

IDEA and other IDEs reuse ProjectConnection instances between build invocations
(i.e. running tasks, executing tests, etc.) to save resources by avoiding
shutting down and re-creating daemon processes. The internal implementation
for ProjectConnection caches the configuration of the target Gradle distribution
when invoking a build for the first time.

This PR fixes the issue by re-reading the gradle-wrapper.properties file and
creating a new internal connection to the target Gradle version if there is a
change in the properties.