Bring back Maven dependencies into Gradle distro
Back in 7.0, we removed most Maven artifacts from the Gradle distro
so they could be dynamically fetched in gradle init --type pom,
in an attempt to reduce the Gradle payload. However, that breaks users
that are behind corporate firewalls.

This commit brings those dependencies back, while moving to a more
recent versions of Maven artifacts.